윤형빈 : 오류 안뜨게

+++ GameEngineContents/GlobalContentsValue.h
@@ -46,7 +46,7 @@ public:
 	static int RemainTwitnum;
 	//bgm
 	static std::string BgmName;
-	static int BgmSpeed;
+	static float BgmSpeed;
 	static bool BgmStop;
 	//체크하고 있는지
 	static bool Check;
