+++ GameEngineContents/StreamAnimation.cpp
@@ -59,7 +59,7 @@ void StreamAnimation::Update(float _DeltaTime)
 		{
 			first = false;
 		}
-		ChangeZPos(std::stoi(Stream::Inst_->StreamMainCol->GetNameConstRef()) - 1);
+		ChangeZPos(static_cast<float>(std::stoi(Stream::Inst_->StreamMainCol->GetNameConstRef())) - 1.0f);
 		Contents = ContentsMap[GlobalContentsValue::Contents];
 		ContentsUpdate(Contents);
 	}
